Crate prctl

Module provides safe abstraction over the prctl interface. Provided functions map to a single prctl() call, although some of them may be usable only on a specific architecture or only with root privileges. All known enums that may be used as parameters are provided in this crate.

Each function provides result which will be Err(errno) in case the prctl() call fails.

To run tests requiring root privileges, enable feature "root_test".

Enums

PrctlCap

PrctlEndian

PrctlFpemu

PrctlMM

PrctlMceKill

PrctlSeccomp

PrctlSecurebits

PrctlTiming

PrctlTsc

PrctlUnalign

Functions

clear_mce_kill

disable_perf_events

drop_capability

enable_perf_events

get_child_subreaper

get_death_signal

get_dumpable

get_endian

get_fpemu

get_keep_capabilities

get_mce_kill

get_name

get_no_new_privileges

get_seccomp

get_securebits

get_thp_disable

get_timer_slack

get_timing

get_tsc

get_unaligned_access

read_capability

set_child_subreaper

set_death_signal

set_dumpable

set_endian

set_fpemu

set_keep_capabilities

set_mce_kill

set_mm

set_name

set_no_new_privileges

set_seccomp_strict

set_securebits

set_thp_disable

set_timer_slack

set_timing

set_tsc

set_unaligned_access